Ensure units & cleanup
* Add core pairs
* Remove old notifyReward script
* proper units for numbers
