+++ scripts/notifyRewardAmount.js
@@ -1,93 +1,79 @@
 // ES5 style
-const config = require("./deploymentConfig");
-const thorify = require("thorify").thorify;
-const Web3 = require("web3");
-const Multirewards = require(config.pathToMultirewardsJson);
-const IERC20 = require(config.pathToIERC20Json);
-const assert = require('assert');
+const find = require('lodash/find');
 const readlineSync = require('readline-sync');
+const ethers = require('ethers');
 
-let network = null;
-let multirewardsAddress = null;
-let rewardTokenAddress = null;
-let rewardAmount = null;
+const Multirewards = require('../build/contracts/MultiRewards.json');
+const IERC20 = require('../build/contracts/IERC20.json');
 
-if (process.argv.length < 6)
-{
+const { getConnex } = require('./utils');
+
+const notifyRewardAmount = async ({
+  connex,
+  multiRewards,
+  network,
+  rewardAmount,
+  rewardToken,
+}) => {
+  console.log(rewardAmount)
+  const approveABI = find(IERC20.abi, { name: 'approve' });
+  const approveMethod = connex.thor.account(rewardToken).method(approveABI);
+
+  const notifyRewardAmountABI = find(Multirewards.abi, { name: 'notifyRewardAmount' });
+  const notifyRewardAmountMethod = connex.thor.account(multiRewards).method(notifyRewardAmountABI);
+
+  const approveClause = approveMethod.asClause(multiRewards, rewardAmount);
+  const notifyRewardAmountClause = notifyRewardAmountMethod.asClause(rewardToken, rewardAmount);
+
+  if (require.main === module && network === 'mainnet') {
+    const input = readlineSync.question("Confirm you want to execute this on the MAINNET? (y/n) ");
+
+    if (input != 'y') process.exit(1);
+  }
+
+  return new Promise(async (resolve, reject) => {
+    try {
+      const approve = await connex.vendor.sign('tx', [approveClause]).request();
+      const notifyRewardAmount = await connex.vendor.sign('tx', [notifyRewardAmountClause]).dependsOn(approve.txid).request();
+      const transaction = await connex.thor.transaction(notifyRewardAmount.txid).getReceipt()
+
+      if (transaction.reverted) {
+        console.log("tx was unsuccessful");
+        reject('tx was unsuccessful')
+      } else {
+        console.log("Set rewards duration was succcessful");
+        resolve('Set rewards duration was succcessful')
+      }
+    } catch(error) {
+      console.error(error);
+      reject(error);
+    }
+  })
+}
+
+// if called directly (from terminal)
+if (require.main === module) {
+  const [network, multiRewards, rewardToken, rewardAmount] = process.argv.slice(2);
+
+  if (!network || !multiRewards || !rewardToken || !rewardAmount) {
     console.error("Usage: node scripts/notifyRewardAmount [mainnet|testnet] [Multirewards address] [Reward token address] [Reward amount (excluding 18 decimals)]");
-    process.exit(1);
-}
-else
-{
-    network = config.network[process.argv[2]];
-    if (network === undefined) {
-        console.error("Invalid network specified");
-        process.exit(1);
-    }
 
-    multirewardsAddress = process.argv[3];
-    rewardTokenAddress = process.argv[4];
-    rewardAmount = Web3.utils.toWei(process.argv[5]);
-}
+    process.exit();
+  }
 
-const web3 = thorify(new Web3(), network.rpcUrl);
-web3.eth.accounts.wallet.add(config.privateKey);
+  (async() => {
+    const connex = await getConnex(network);
 
-notifyRewardAmount = async() =>
-{
-    // This is the address associated with the private key
-    const walletAddress = web3.eth.accounts.wallet[0].address;
+    const result = await notifyRewardAmount({
+      connex,
+      multiRewards,
+      network,
+      rewardAmount,
+      rewardToken,
+    });
 
-    console.log("Using wallet address:", walletAddress);
-    console.log("Using RPC:", web3.eth.currentProvider.RESTHost);
-
-    try
-    {
-        let transactionReceipt = null;
-        const multirewardsContract = new web3.eth.Contract(Multirewards.abi, multirewardsAddress);
-        const rewardTokenContract = new web3.eth.Contract(IERC20.abi, rewardTokenAddress);
-
-        console.log("Attempting ERC20 approve for transfer, amount:", rewardAmount);
-
-        if (network.name == "mainnet")
-        {
-            let input = readlineSync.question("Confirm you want to execute this on the MAINNET? (y/n) ");
-            if (input != 'y') process.exit(1);
-        }
-
-        await rewardTokenContract
-                .methods
-                .approve(multirewardsAddress,
-                         rewardAmount)
-                .send({ from: walletAddress })
-                .on("receipt", (receipt) => {
-                    transactionReceipt = receipt;
-                });
-
-        console.log("Approve successful, tx hash:", transactionReceipt.transactionHash);
-        console.log("Attempting notifyRewardAmount for multireward address:", multirewardsAddress);
-        console.log("For staking token:", await multirewardsContract
-                                            .methods
-                                            .stakingToken()
-                                            .call());
-
-        await multirewardsContract
-                .methods
-                .notifyRewardAmount(
-                    rewardTokenAddress,
-                    rewardAmount
-                )
-        .send({ from: walletAddress })
-        .on("receipt", (receipt) => {
-            transactionReceipt = receipt;
-        });
-
-        console.log("Transaction Hash:", transactionReceipt.transactionHash);
-    }
-    catch(error)
-    {
-        console.log("Execution failed with:", error)
-    }
+    console.log(result)
+  })();
 }
 
-notifyRewardAmount();
+module.exports = notifyRewardAmount;
